requirements.txt
boto==2.33.0
invoke==0.9.0
PyYAML==3.11
Jinja2==2.7.3
